Release v0.9.171 / v1.0.0-rc166 / nodus v0.10.5 [BUILD] [RELEASE]
Kyber1024 post-quantum channel encryption — 3 layers:
  1. Client↔Server transport (all T2 traffic encrypted)
  2. Inter-node transport (replication, heartbeat encrypted)
  3. Circuit E2E (onion layer, relay-blind)

Fixes since v0.10.4:
- Inter-node replication routed through encrypt path
- Auth gate kept closed during Kyber handshake
- Frame loop break on crypto state change
- In-flight plaintext frames skipped gracefully
- Android format warning fix (%zu → %u cast)

Security: H-01 (transport encryption), H-02, H-03 FIXED.
Tests: 37/37 nodus, 10 channel crypto, 12/12 circuit live.
Production: 7/7 cluster encrypted, punk↔chip messaging verified.

+++ nodus/src/transport/nodus_tcp.c
@@ -220,8 +220,8 @@ static bool try_parse_frames(nodus_tcp_t *tcp, nodus_tcp_conn_t *conn) {
                          * silently (periodic data will be re-sent). Only
                          * disconnect on repeated failures (real attack). */
                         cc->rx_counter = 0;  /* Reset counter for next valid frame */
-                        QGP_LOG_WARN(LOG_TAG_TCP, "decrypt skip: conn=%s:%d frame_len=%zu (in-flight plaintext)",
-                                     conn->ip, conn->port, frame.payload_len);
+                        QGP_LOG_WARN(LOG_TAG_TCP, "decrypt skip: conn=%s:%d frame_len=%u (in-flight plaintext)",
+                                     conn->ip, conn->port, (unsigned)frame.payload_len);
                         free(dec_buf);
                         dec_buf = NULL;
                         /* Skip this frame, continue processing */
